Install qemu and create agl-egvirt machine feature
qemu is the userspace emulator needed to run KVM virtual machines.
This patch adds a bbappend for qemu in the meta-egvirt layer, and
includes it in the packages to be installed in the AGL distro.
To do this, the new agl-egvirt machine feature has been created
and is used to enable virtualization specific extensions
created by the EG-VIRT group for AGL.

In order to reduce its impact on the AGL image, the qemu targets
have been limited to aarch64 and x86_64 only.

Changelog v2:
- agl-egvirt added to MACHINE_FEATURES
- use of class_target to not create conflict by the QEMU version
  used in the SDK and the one used by KVM
- README file updated accordingly

Bug-AGL: SPEC-718

Change-Id: I27288382443ba2bfe14d574ade2e691a63da145d
Signed-off-by: Michele Paolino <m.paolino@virtualopensystems.com>