Fix 4a-mixer parsing error on nogfx build variant
The nogfx build variant fails with 4a enabled ( https://gerrit.automotivelinux.org/gerrit/#/c/15051/ )
This is documented in SPEC-1588.
To fix this we apply a conditional inclusion in meta-audio-4a-framework to only pull-in the mixer if
we are including meta-agl-demo. Can be more fine-grained (profile-graphical-qt5 maybe) later.

Bug-AGL: SPEC-1588

Fast inclusion requested as it is a blocker for RC2 due to enabling 4a in CI.

Change-Id: I59b82748ba8e1736358318373d29f8f9fed2a5c5
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>