set LIBASOUND_THREAD_SAFE=0 for 4a
Since the integration of bluez-alsa support in softmixer,
and as warned by the readme of bluez-alsa, there are potential
(and effective !) deadlocks in libasound. These deadlocks come
especially when uing ioplug PCMs.
This hack disables the calls to snd_pcm_lock/unlock in libasound.

Notice that the design that consists in using such PCM plugs
for communicating with the bluealsa daemon is just a short term
solution. When the direct support of bluezalsa communication
is integrated in the softmixer, that hack should be removed.

Change-Id: Ib6376fd2e145b3240ab1a68b9c760ff9e4c063c4
Signed-off-by: Thierry Bultel <thierry.bultel@iot.bzh>