Move hal configuration files into own repository
... and own recipe

This allows to manage the various types of configuration files
independently of the source code as these might not always fit
into the git repo context-wise.

Initial recipe is 4a-hal-device-config.bb .
This is re-using the same mechanism as previously used.

Samples for alternative implementations and vendor hals are
included as well to explore improved ways of handling
the multipe configuration files possible.

Bug-AGL: SPEC-2011

Change-Id: Ib04b5c4e7b764bed85dbceb50a62a8c6fd031c5a
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>