weston: Add patch to improve gst-record robustness
Add a patch to do some retries with a delay when the gst-record
feature is initializing and attempts to query the omx encoder plugin
for its buffer pools.  It is not clear why exactly the encoder is
not consistently ready at this point, but in my testing this change
greatly improved the robustness of weston startup with gst-record
enabled.

Change-Id: Ibd26ad64be035e1f9d40cc3cd9337e9efeabfad4
Signed-off-by: Scott Murray <scott.murray@konsulko.com>