bluez-alsa: removed the PCM hack, and build a shared lib
This removes the deprecated PCM proxy hack, and brings mandatory
features for the softmixer.
This change applies on top of the latest bluez-alsa master of 2018-12-04
and brings 3 patches:
1) builds a shared library and links bluez-alsa and ioplug controls & pcm
to it.
2) func log by default
3) increase the number of max connections from clients (pcm & ctrls)
-----
Upstream status for the patches:
1) pending. Based on a submitted PR that is not clean enough and
that the maintainer does not want to take before at least 6 months
https://github.com/Arkq/bluez-alsa/pull/95

2) pending. Was submitted as part of the Ofono support in bluez-alsa,
but was not retained. See the original PR for Ofono for that.

3) pending. The maintainer is OK with this but wants to refactor the
client-server protocol in the coming months. In a private mail of
12/06/2018:

"Yes, you can increase this hardcoded limit as a temporary workaround.
I've put in there (and made it rather low) as a reminder, that clients
are stored in the list, which might not be optimal for larger number of
connected clients. In the future bluealsa will support audio mixing, so
connecting many clients to a single transport will be a common case. So,
I've thought about incorporating more efficient structure for storing
(in particular searching) connected client - something with log(n)
search complexity. Then, it might be worth to implement this container
to be dynamic with an upper capacity limit (which might be high, like
200-300 clients)."

Change-Id: Ia771613a7a83cc9ed6108399cdaacdeda603d39b
Signed-off-by: Thierry Bultel <thierry.bultel@iot.bzh>