agl-service-windowmanager: add weston-ready systemd dependency
Add installation of systemd over-ride configuration that adds a
dependency on weston-ready to ensure that the windowmanager and
its dependencies start after weston is actually initialized.
In testing, this seems to resolve the issue of homescreen and
launcher failing to start in situations such as when using MOST
audio and/or the gst-record feature in Weston.

Bug-AGL: SPEC-2115

Change-Id: Id15045357f30d3f447037026d0ab960ae2569d69
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
(cherry picked from commit 7650fff17bd502e80feb28dc9bf5927bc5ecf52c)
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>