meta-pipewire: add patch to security-manager for pipewire
This grants permission to all applications to be able to communicate
with pipewire. This is a temporary measure that we need to rethink
in the future.

This commit also changes the priority of the meta-pipewire layer
because the security-manager patch needs to be applied after the
patch to the same file from meta-app-framework, which has priority 70.

Bug-AGL: SPEC-2554

Change-Id: I9df47ea1a40e1fa97900bb5dbc21bc11765eb7ab
Signed-off-by: George Kiagiadakis <george.kiagiadakis@collabora.com>