meta-html5-framework/wam: adjust how wam is started
WebAppMgr has to be launched for each user (but not through
the systemd user session as it would require too much privileges for the user).

For this purpose, we create a systemd service instanciated for each user
by the application framework through afm-user-session@ service.

This patch also distributes the environment and service file near the recipe
instead of having those integration files in the sources (it's easier to follow
AGL evolution without changing the upstream source code)

Also, for the HTML5 demos to work, each app must request the following permissions in
their config.xml:
	<param name="urn:AGL:permission::public:display" value="required" />
	<param name="urn:AGL:permission::public:audio" value="required" />

Bug-AGL: SPEC-2586

Change-Id: I4286fd9e1d1a1dcff48d3fdbfab2bf8d57f0fa0e
Signed-off-by: Stephane Desneux <stephane.desneux@iot.bzh>