pipewire config: enable bluez modules in pipewire and wireplumber
In pipewire we only enable that if the bluez5 feature is enabled
because otherwise the module is not compiled and will error out
at runtime.

In wireplumber we always enable it since it does not depend on bluez.

Bug-AGL: SPEC-2792

Change-Id: I53bbc387bccd953633affd4fc34949a2fd488ed7
Signed-off-by: George Kiagiadakis <george.kiagiadakis@collabora.com>
(cherry picked from commit 9fc6d8e9bbc9746777c03471b643992abcb182a5)