audiomixer: temporarily disable code in audiomixer to make the build work
This essentially removes all mixer controls from the mixer, while it still
allows the service to build and run, so that the rest of the system is not
broken.

The API used by audiomixer (the endpoint extension) has changed and we
are still working on porting all the pieces involved.

Bug-AGL: SPEC-2837

Change-Id: I12879f13f545c8cebe913860801cc61ed921e009
Signed-off-by: George Kiagiadakis <george.kiagiadakis@collabora.com>