alexa-voiceagent-service: update config and database file paths
Add patch to alexa-voiceagent-service to update the default paths
used for the database and configuration files to move them out of the
binding installation hierarchy. This avoids the permissions problems
stemming from the new security model of running as non-root. Also
reworked the main configuration JSON file location logic to check for
the file in /etc/xdg/AGL and then in AFM_WORKDIR (app-data directory),
before falling back to the original location in var/config under the
binding installation directory.

Bug-AGL: SPEC-2845

Change-Id: Id7636435163005148a6291a56a1ada9896c0b3cc
Signed-off-by: Scott Murray <scott.murray@konsulko.com>