meta-speech-framework: rework voiceagent configuration
After agl-service-voice-high rework to make the voiceagent(s) used
externally configurable, replace the previous virtual/voiceagent
PREFERRED_RPROVIDER scheme with a virtual/voice-high-config that
just selects the appropriate configuration package based on the
specified feature. If agl-speech-framework is specified on its own,
a default configuration file specifying no default voiceagent will
be installed. If agl-voiceagent-alexa is specified, the default
configuration will specify the Alexa voiceagent as the default.
Recipes have been added for the new configuration file packages,
and the alexa-voiceagent-service recipe has been tweaked to create
a package with the Alexa per-voiceagent configuration file and add
it as a dependency.

Bug-AGL: SPEC-2898

Change-Id: Ib6e3952d04df5795d94bb38cd314b0c79e5b60e5
Signed-off-by: Scott Murray <scott.murray@konsulko.com>