meta-html5-framework: define html-only image
The layer now defines a standalone image agl-image-graphical-html5,
containing only the minimum packages to run the web application
manager (WAM). No webapps are included.

It redefines packagegroups to use the more standard names
agl-profile-graphical-html5 and agl-appfw-html5.

It also adds a dependency on agl-profile-graphical, so we don't
require users to add other features explicitly.

The .bbappends for agl-demo-platform* images have been removed, so the
layer doesn't depend on meta-agl-demo. We will provide image
definitions for WAM + demo apps in meta-agl-demo in a later commit.

To sum up, an HTML-only image can be built with:

  aglsetup.sh -m $MACHINE agl-html5-framework
  bitbake agl-image-graphical-html5

To add WAM to an agl-demo-platform image, both agl-demo and
agl-html5-framework features must be enabled, and this must be added
to local.conf or local.dev.inc:

  IMAGE_INSTALL_append = " packagegroup-agl-appfw-html5"

Bug-AGL: SPEC-2322

Change-Id: I648dd9f509faf77bf0c8520c9773de073865b90e
Signed-off-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>