meta-aac: Rework agl-voiceagent-alexa packaging
Rather than appending the agl-voiceagent-alexa packages onto the
existing packagegroup-agl-core-services, instead define a new specific
packagegroup-agl-voiceagent-alexa, and an associated image feature
that is used to include it into an image with IMAGE_FEATURES. This
seems better for the longer-term goal of a binary package feed, and
IMAGE_FEATURES can be used by other recipes in a similar manner as
DISTRO_FEATURES, but with a lot less build dependency impact if
changed. This will be used in a parallel change to meta-agl-demo to
pull in the new alexa-viewer application when using the Alexa
voiceagent.

Bug-AGL: SPEC-3110

Change-Id: I15fb07560fef703e3471bce9e50596cb4a6667b4
Signed-off-by: Scott Murray <scott.murray@konsulko.com>