Remove agl-gstrecorder / meta-gstrecorder-rcar-gen3
Remove the agl-gstrecorder feature and its meta-gstrecorder-rcar-gen3
layer. The feature itself is going to be replaced with a more generic
agl-weston-remoting feature in meta-agl. From the layer, the remoting
gbm fix is no longer required for rcar3 in Weston 8.0.0, and the
remoting initialization timing patch is generic, and is thus going to
be applied for all platforms in meta-agl-profile-graphical to make
cluster support generic.

Bug-AGL: SPEC-3395

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I69ea73cc14a30e084fb0443628956fdc2572060e