meta-agl-jailhouse: Add support for RPi4 with 2G/4G/8G of memory
The Jailhouse configuration files had to be modified to append the additional
memory regions present on the higher mem variants of the RPi4.

To avoid the use of mem= kernel parameter to fix the memory that can be used
by the root cell, a device tree overlay with reserved-memory nodes is used
instead.

To avoid conflict with the GPU memory, the memory region used for Jailhouse
and the non-root cells was moved from 0x30000000 to 0x20000000.

Bug-AGL: SPEC-3507

Signed-off-by: Jakub Luzny <jakub@luzny.cz>
Change-Id: Iabd67504c3df0349dd62e45f12d1019de14328f1