libweston: Don't request repaint in output enable
Allow for the compositor to start without first displaying
a blank screen when no startup transition is used.

This patch is under review in the mainline weston repository.
It should be useable for other use cases besides drm-lease,
but right now this patch causes the weston test suite to fail
by unconvering an unrelated problem with wayland event handling[1].

For now, keep this with the drm-lease support under agl-devel until
the weston test suite issue is resolved and the patch is merged upstream.

[1]: https://gitlab.freedesktop.org/wayland/wayland/-/issues/201

Bug-AGL: SPEC-3816

Change-Id: Ib5a2f0dde833d88a17bd4e659ff1acee81acd21f
Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>