meta-agl-lxc: Fix lack of gles modules in demo image
With current integration of lxc-host-image-demo for rcar-gen3,
weston fails to start due to lack of gles modules so that demo
application doesn't come up. This change fixes it with adding
kernel-module-gles into lxc-host-image-demo image, only for
rcar-gen3 boards as of this moment (since lxc-host-image-demo is
maintained only for h3/m3 ulcb boards currently).

Signed-off-by: Hiroyuki Ishii <ishii.hiroyuki002@jp.panasonic.com>
Change-Id: I7874b125e6937c43bc706f63c971cb33d1a47a6c