meta-agl-jailhouse: Fix memmap command line parameter
Jailhouse requires a contiguous piece of RAM for itself and additional
cells. This has to be preallocated during emulation of the image and
is typically done by adding `memmap=<size>$<memory_address>`, and here
escape sequencing before the dollar should only be added  when there is
a need to put this parameter in GRUB2 variables in /etc/default/grub.

Fix it by removing `\` escape sequence from memmap command line parameter
of the virtual machine's kernel.

Signed-off-by: Anmol <anmol.karan123@gmail.com>
Change-Id: I9981a359d1cb00900153e7692c6c1175cb0a3ce6