Convert to new override syntax
This is mostly the result of running a slightly customized version
of the convert-overrides.py script from poky with additional
overrides added.  A few minor fixups were done by hand afterwards
during a review of the changes.

The intent of these changes is to minimize the effort to keep the
"next" branch that builds against poky master up to date and tested
in preparation for the switch to the next Yocto LTS release in
early 2022.

Bug-AGL: SPEC-4052

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I0bbae97331abee1367b368ce54ab39e3162a12ec