meta-agl-lxc: Support rcar3 kernel module headers in guests
Add a bbclass that can be used to disable compilation in
module.bbclass based recipes, and dynamically activated bbappends for
the rcar3 kernel module recipes that use it and guest-build-specific
do_install overrides to get the userspace API headers installed
transparently to the userspace recipes.  This avoids the complications
that would arise from using a full linux-renesas kernel build for
guest container images.

Bug-AGL: SPEC-4089

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I917e66ca024a2a507e615ba3a2a58c603a691142